燃尽了

+++ linear_module.py
@@ -6,7 +6,7 @@ import torch.nn.init as init
 class Linear(nn.Module):
     def __init__(self, in_features, out_features, device = None, dtype = None):
         super().__init__()
-        self.W = torch.empty(out_features, in_features, device=device,dtype =dtype)
+        self.W = torch.empty(out_features, in_features, device=device,dtype = dtype)
         self.W = nn.Parameter(self.W)
         std = math.sqrt(2.0 / (in_features + out_features))
         init.trunc_normal_(self.W, mean=0.0, std=std, a= -3*std,b= 3*std )
